Moved nbtPath
