+++ src/main/java/fr/estecka/variantscit/modules/ComponentDataModule.java
@@ -3,8 +3,8 @@ package fr.estecka.variantscit.modules;
 import com.mojang.serialization.DataResult;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import fr.estecka.variantscit.NbtPath;
 import fr.estecka.variantscit.VariantsCitMod;
+import fr.estecka.variantscit.nbt.NbtPath;
 import net.minecraft.component.ComponentType;
 import net.minecraft.nbt.AbstractNbtNumber;
 import net.minecraft.nbt.NbtElement;
